Create some tests for chunkedUntil
