+++ jmdict/src/commonMain/kotlin/com/boswelja/jmdict/JmDict.kt
@@ -21,21 +21,6 @@ expect class JmDictReader {
     suspend fun streamJmDict(): Sequence<Entry>
 }
 
-internal fun <T> Sequence<T>.chunkedUntil(predicate: (T) -> Boolean): Sequence<List<T>> {
-    return sequence {
-        var list = mutableListOf<T>()
-        this@chunkedUntil.forEach {
-            if (!predicate(it)) {
-                list.add(it)
-            } else {
-                yield(list)
-                list = mutableListOf(it)
-            }
-        }
-        yield(list)
-    }
-}
-
 internal fun Sequence<String>.asEntrySequence(): Sequence<Entry> {
     return this
         .dropWhile { !it.contains("<entry>") }
