+++ jmdict/src/commonTest/kotlin/com/boswelja/jmdict/SequenceExtTest.kt
@@ -1,0 +1,35 @@
+package com.boswelja.jmdict
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertTrue
+
+class SequenceExtTest {
+
+    @Test
+    fun `chunkedUntil with empty sequence`() {
+        val emptySequence = emptySequence<String>()
+        val chunkedUntil = emptySequence.chunkedUntil { it == "Hello, world!" }
+        assertTrue(chunkedUntil.none())
+    }
+
+    @Test
+    fun `chunkedUntil where no elements match predicate`() {
+        val targetSequence = sequenceOf(
+            "Hello,",
+            "World!",
+            "I'm",
+            "a",
+            "sequence!"
+        )
+        val chunkedUntil = targetSequence.chunkedUntil { it.endsWith("~") } // Nothing in targetSequence ends with ~
+        assertEquals(
+            1,
+            chunkedUntil.count()
+        )
+        assertEquals(
+            targetSequence.toList(),
+            chunkedUntil.first()
+        )
+    }
+}
